+++ src/client/index.ts
@@ -1,5 +1,5 @@
 export * from "./soda-client";
 export * from "./soda-context";
-export * from "./context-soda-host-decorator";
+export * from "./soda-host-decorator";
 export * from "./soda-resource-id";
 export * from "./soda-resource";
